fix minor issue with decals,EMreels and textboxes setting the defaults to registry and truncate stored floats to what is read back again
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@1173 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ dispreel.h
@@ -19,7 +19,7 @@ public:
     char        m_szImage[MAXTOKEN];    // image to use for the decals.
     ReelType    m_reeltype;
 	BOOL		m_fUseImageGrid;
-    long		m_imagesPerGridRow;
+    int			m_imagesPerGridRow;
     int			m_reelcount;			// number of individual reel in the set
     float       m_width, m_height;      // size of each reel
     float       m_reelspacing;          // spacing between each reel and the boarders
@@ -27,7 +27,7 @@ public:
 	int			m_digitrange;			// max number of digits per reel (usually 9)
 
     char        m_szSound[MAXTOKEN];    // sound to play for each turn of a digit
-    long        m_updateinterval;       // time in ms between each animation update
+    int	        m_updateinterval;       // time in ms between each animation update
 
     COLORREF    m_backcolor;            // colour of the background
     COLORREF    m_reelcolor;            // colour of the reels (valid if m_reeltype = ReelText)
